v0.4 b039 +ADD MembersCountChannelListener.    - Is changing the name of the channel in the count of the current members on the discord. + ADD NSFW System. + ADD SaveCurrentTimesCommand (a developer command)
+ HOTFIX If a user has the regular customer role, his active time was not updated.

+++ src/main/java/xyz/jupp/discord/commands/handler/CommandHandler.java
@@ -31,7 +31,9 @@ public class CommandHandler extends ListenerAdapter {
             for (Command command : getCommands()) {
                 if (command.getCommand().startsWith(args[0])) {
                     command.action(args, event);
-                    event.getMessage().delete().queue();
+                    if (!event.getAuthor().isBot()){
+                        event.getMessage().delete().queue();
+                    }
                 }
             }
